fixed dependecies and configured routers into one
