+++ api/server.js
@@ -2,9 +2,7 @@ const express = require('express');
 const cors = require('cors');
 const helmet = require('helmet');
 
-const authenticate = require('../auth/authenticate-middleware.js');
-const authRouter = require('../auth/auth-router.js');
-const jokesRouter = require('../jokes/jokes-router.js');
+const configureRoutes = require('../config/routes.js');
 
 const server = express();
 
@@ -12,7 +10,6 @@ server.use(helmet());
 server.use(cors());
 server.use(express.json());
 
-server.use('/api/auth', authRouter);
-server.use('/api/jokes', authenticate, jokesRouter);
+configureRoutes(server);
 
 module.exports = server;
