+++ config/routes.js
@@ -1,13 +1,17 @@
-const router = require('express').Router();
-
+const axios = require('axios');
 const bcrypt = require('bcryptjs');
+const jwt = require('jsonwebtoken');
+
+const { authenticate, jwtKey } = require('../auth/authenticate-middleware.js');
 const Users = require('../users/users-model.js');
-const jwtKey = require('../auth/authenticate-middleware.js');
 
+module.exports = server => {
+  server.post('/api/register', register);
+  server.post('/api/login', login);
+  server.get('/api/jokes', authenticate, getJokes);
+};
 
-
-router.post('/register', (req, res) => {
-  // implement registration
+function register(req, res) {
   let user = req.body;
   const hash = bcrypt.hashSync(user.password, 10)
   user.password = hash;
@@ -23,10 +27,9 @@ router.post('/register', (req, res) => {
         .status(500)
         .json(error);
     });
-});
+};
 
-router.post('/login', (req, res) => {
-  // implement login
+function login(req, res) {
   const { username, password } = req.body;
 
   Users.findBy({ username })
@@ -58,20 +61,32 @@ router.post('/login', (req, res) => {
           errorMessage: 'login error' 
         });
     }); 
+}
 
-});
+function getJokes(req, res) {
+  const requestOptions = {
+    headers: { accept: 'application/json' },
+  };
+
+  axios
+    .get('https://icanhazdadjoke.com/search', requestOptions)
+    .then(response => {
+      res.status(200).json(response.data.results);
+    })
+    .catch(err => {
+      res.status(500).json({ message: 'Error Fetching Jokes', error: err });
+    });
+}
 
 function generateToken(user) {
   const payload = {
     subject: user.id,
     username: user.username
   };
-
+  
   const options = {
     expiresIn: '1h'
   };
-
+  
   return jwt.sign(payload, jwtKey, options);
 }
-
-module.exports = router;
