+++ package.json
@@ -24,6 +24,7 @@
     "dotenv": "^8.1.0",
     "express": "^4.17.1",
     "helmet": "^3.20.0",
+    "jsonwebtoken": "^8.5.1",
     "knex": "^0.19.2",
     "sqlite3": "^4.1.0"
   },
